# openapi.json
{
  "components": {
    "schemas": {
      "AuditRecord": {
        "properties": {
          "audit_ref": {
            "title": "Audit Ref",
            "type": "string"
          },
          "chain_valid": {
            "title": "Chain Valid",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "situation_id": {
            "title": "Situation Id",
            "type": "string"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/AuditStep"
            },
            "title": "Steps",
            "type": "array"
          }
        },
        "required": [
          "audit_ref",
          "situation_id",
          "created_at",
          "steps",
          "chain_valid",
          "model_id",
          "provider"
        ],
        "title": "AuditRecord",
        "type": "object"
      },
      "AuditStep": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "duration_ms": {
            "minimum": 0.0,
            "title": "Duration Ms",
            "type": "number"
          },
          "entry_hash": {
            "title": "Entry Hash",
            "type": "string"
          },
          "inputs": {
            "additionalProperties": true,
            "default": {},
            "title": "Inputs",
            "type": "object"
          },
          "outputs": {
            "additionalProperties": true,
            "default": {},
            "title": "Outputs",
            "type": "object"
          },
          "prev_hash": {
            "title": "Prev Hash",
            "type": "string"
          },
          "seq": {
            "title": "Seq",
            "type": "integer"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "step": {
            "title": "Step",
            "type": "string"
          },
          "tier": {
            "title": "Tier",
            "type": "string"
          }
        },
        "required": [
          "seq",
          "step",
          "tier",
          "detail",
          "started_at",
          "duration_ms",
          "entry_hash",
          "prev_hash"
        ],
        "title": "AuditStep",
        "type": "object"
      },
...
      "CrewMember": {
        "properties": {
          "circadian_offset_h": {
            "default": 0.0,
            "title": "Circadian Offset H",
            "type": "number"
          },
          "duty_log": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/DutyEntry"
            },
            "title": "Duty Log",
            "type": "array"
          },
          "id": {
            "examples": [
              "CM-01"
            ],
            "title": "Id",
            "type": "string"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "qualified_for": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Qualified For",
            "type": "array"
          },
          "role": {
            "examples": [
              "commander"
            ],
            "title": "Role",
            "type": "string"
          },
          "sleep_log": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SleepEntry"
            },
            "title": "Sleep Log",
            "type": "array"
          }
        },
        "required": [
          "id",
          "role"
        ],
        "title": "CrewMember",
        "type": "object"
      },
      "CrewReadiness": {
        "description": "Zone 1 payload: per-crew state at the moment of evaluation.",
        "properties": {
          "alertness_score": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Alertness Score",
            "type": "number"
          },
          "asleep_at_window_start": {
            "title": "Asleep At Window Start",
            "type": "boolean"
          },
          "baseline_alertness": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Baseline Alertness",
            "type": "number"
          },
          "confidence": {
            "enum": [
              "high",
              "moderate",
              "low",
              "insufficient"
            ],
            "title": "Confidence",
            "type": "string"
          },
          "crew_member": {
            "title": "Crew Member",
            "type": "string"
          },
          "curve": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/CurvePoint"
            },
            "title": "Curve",
            "type": "array"
          },
          "data_coverage": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Data Coverage",
            "type": "number"
          },
          "delta_vs_baseline": {
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Delta Vs Baseline",
            "type": "number"
          },
          "hours_awake": {
            "minimum": 0.0,
            "title": "Hours Awake",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "reference_at": {
            "format": "date-time",
            "title": "Reference At",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "sleep_debt_h": {
            "minimum": 0.0,
            "title": "Sleep Debt H",
            "type": "number"
          },
          "status": {
            "enum": [
              "nominal",
              "watch",
              "degraded"
            ],
            "title": "Status",
            "type": "string"
          },
          "trend": {
            "enum": [
              "improving",
              "stable",
              "declining"
            ],
            "title": "Trend",
            "type": "string"
          },
          "window_min_alertness": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Window Min Alertness",
            "type": "number"
          },
          "window_min_at": {
            "format": "date-time",
            "title": "Window Min At",
            "type": "string"
          },
          "workload_score": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Workload Score",
            "type": "number"
          }
        },
        "required": [
          "crew_member",
          "name",
          "role",
          "reference_at",
          "asleep_at_window_start",
          "alertness_score",
          "baseline_alertness",
          "delta_vs_baseline",
          "workload_score",
          "sleep_debt_h",
          "hours_awake",
          "window_min_alertness",
          "window_min_at",
          "data_coverage",
          "confidence",
          "status",
          "trend"
        ],
        "title": "CrewReadiness",
        "type": "object"
      },
...
      "DecisionRecord": {
        "properties": {
          "audit_ref": {
            "title": "Audit Ref",
            "type": "string"
          },
          "decision": {
            "title": "Decision",
            "type": "string"
          },
          "decision_id": {
            "title": "Decision Id",
            "type": "string"
          },
          "operator": {
            "title": "Operator",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "recorded_at": {
            "format": "date-time",
            "title": "Recorded At",
            "type": "string"
          },
          "situation_id": {
            "title": "Situation Id",
            "type": "string"
          }
        },
        "required": [
          "decision_id",
          "situation_id",
          "audit_ref",
          "decision",
          "operator",
          "reason",
          "recorded_at"
        ],
        "title": "DecisionRecord",
        "type": "object"
      },
      "DecisionRequest": {
        "properties": {
          "audit_ref": {
            "title": "Audit Ref",
            "type": "string"
          },
          "decision": {
            "enum": [
              "approved",
              "overridden"
            ],
            "title": "Decision",
            "type": "string"
          },
          "operator": {
            "default": "flight_surgeon_console",
            "title": "Operator",
            "type": "string"
          },
          "reason": {
            "default": "",
            "title": "Reason",
            "type": "string"
          },
          "situation_id": {
            "title": "Situation Id",
            "type": "string"
          }
        },
        "required": [
          "situation_id",
          "audit_ref",
          "decision"
        ],
        "title": "DecisionRequest",
        "type": "object"
      },
...
      "EvaluationRequest": {
        "properties": {
          "crew": {
            "items": {
              "$ref": "#/components/schemas/CrewMember"
            },
            "title": "Crew",
            "type": "array"
          },
          "evaluation_window": {
            "$ref": "#/components/schemas/EvaluationWindow"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/Task"
            },
            "title": "Tasks",
            "type": "array"
          }
        },
        "required": [
          "evaluation_window",
          "crew",
          "tasks"
        ],
        "title": "EvaluationRequest",
        "type": "object"
      },
      "EvaluationResponse": {
        "properties": {
          "archived_low_risk": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Archived Low Risk",
            "type": "array"
          },
          "evaluation_id": {
            "title": "Evaluation Id",
            "type": "string"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "readiness": {
            "items": {
              "$ref": "#/components/schemas/CrewReadiness"
            },
            "title": "Readiness",
            "type": "array"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "scenario_note": {
            "default": "",
            "title": "Scenario Note",
            "type": "string"
          },
          "scenario_title": {
            "default": "",
            "title": "Scenario Title",
            "type": "string"
          },
          "situations": {
            "items": {
              "$ref": "#/components/schemas/Situation"
            },
            "title": "Situations",
            "type": "array"
          },
          "tier_status": {
            "$ref": "#/components/schemas/TierStatus"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineTask"
            },
            "title": "Timeline",
            "type": "array"
          },
          "window": {
            "$ref": "#/components/schemas/EvaluationWindow"
          }
        },
        "required": [
          "evaluation_id",
          "generated_at",
          "window",
          "readiness",
          "timeline",
          "situations",
          "tier_status"
        ],
        "title": "EvaluationResponse",
        "type": "object"
      },
      "EvaluationWindow": {
        "properties": {
          "end": {
            "format": "date-time",
            "title": "End",
            "type": "string"
          },
          "start": {
            "format": "date-time",
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "start",
          "end"
        ],
        "title": "EvaluationWindow",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Situation": {
        "properties": {
          "alertness_score": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Alertness Score",
            "type": "number"
          },
          "audit_ref": {
            "title": "Audit Ref",
            "type": "string"
          },
          "circadian_flag": {
            "title": "Circadian Flag",
            "type": "boolean"
          },
          "confidence": {
            "enum": [
              "high",
              "moderate",
              "low",
              "insufficient"
            ],
            "title": "Confidence",
            "type": "string"
          },
          "corpus_manifest": {
            "default": "",
            "title": "Corpus Manifest",
            "type": "string"
          },
          "crew_member": {
            "title": "Crew Member",
            "type": "string"
          },
          "crew_member_name": {
            "default": "",
            "title": "Crew Member Name",
            "type": "string"
          },
          "crew_role": {
            "default": "",
            "title": "Crew Role",
            "type": "string"
          },
          "evidence": {
            "$ref": "#/components/schemas/Evidence"
          },
          "outcome": {
            "enum": [
              "recommendation",
              "refusal"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "recommendation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Recommendation"
              },
              {
                "type": "null"
              }
            ]
          },
          "refusal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Refusal"
              },
              {
                "type": "null"
              }
            ]
          },
          "risk_level": {
            "enum": [
              "nominal",
              "elevated",
              "critical"
            ],
            "title": "Risk Level",
            "type": "string"
          },
          "situation_id": {
            "title": "Situation Id",
            "type": "string"
          },
          "task": {
            "title": "Task",
            "type": "string"
          },
          "task_criticality": {
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Task Criticality",
            "type": "string"
          },
          "task_label": {
            "default": "",
            "title": "Task Label",
            "type": "string"
          },
          "task_scheduled": {
            "format": "date-time",
            "title": "Task Scheduled",
            "type": "string"
          },
          "task_type": {
            "title": "Task Type",
            "type": "string"
          },
          "workload_score": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Workload Score",
            "type": "number"
          }
        },
        "required": [
          "situation_id",
          "crew_member",
          "task",
          "task_type",
          "task_scheduled",
          "task_criticality",
          "alertness_score",
          "workload_score",
          "circadian_flag",
          "risk_level",
          "confidence",
          "outcome",
          "evidence",
          "audit_ref"
        ],
        "title": "Situation",
        "type": "object"
      },
...
      "Task": {
        "properties": {
          "assigned_to": {
            "title": "Assigned To",
            "type": "string"
          },
          "criticality": {
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Criticality",
            "type": "string"
          },
          "duration_min": {
            "default": 60,
            "title": "Duration Min",
            "type": "integer"
          },
          "id": {
            "examples": [
              "T-119"
            ],
            "title": "Id",
            "type": "string"
          },
          "label": {
            "default": "",
            "title": "Label",
            "type": "string"
          },
          "scheduled": {
            "format": "date-time",
            "title": "Scheduled",
            "type": "string"
          },
          "type": {
            "examples": [
              "orbital_burn"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "criticality",
          "scheduled",
          "assigned_to"
        ],
        "title": "Task",
        "type": "object"
      },
      "TierStatus": {
        "description": "Zone 6 payload: which tier is live, which is mocked, which is degraded.",
        "properties": {
          "corpus_manifest": {
            "default": "",
            "title": "Corpus Manifest",
            "type": "string"
          },
          "degraded": {
            "default": false,
            "title": "Degraded",
            "type": "boolean"
          },
          "degraded_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Degraded Reason"
          },
          "deterministic": {
            "title": "Deterministic",
            "type": "string"
          },
          "orchestration": {
            "title": "Orchestration",
            "type": "string"
          },
          "provider_chain": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Provider Chain",
            "type": "array"
          },
          "reasoning": {
            "title": "Reasoning",
            "type": "string"
          },
          "retrieval": {
            "title": "Retrieval",
            "type": "string"
          },
          "served_by": {
            "default": "",
            "title": "Served By",
            "type": "string"
          }
        },
        "required": [
          "deterministic",
          "retrieval",
          "reasoning",
          "orchestration"
        ],
        "title": "TierStatus",
        "type": "object"
      },
      "TimelineTask": {
        "description": "Zone 2 payload: a task plotted against predicted alertness.",
        "properties": {
          "assigned_name": {
            "title": "Assigned Name",
            "type": "string"
          },
          "assigned_to": {
            "title": "Assigned To",
            "type": "string"
          },
          "circadian_low": {
            "title": "Circadian Low",
            "type": "boolean"
          },
          "criticality": {
            "enum": [
              "low",
              "medium",
              "high"
            ],
            "title": "Criticality",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "predicted_alertness": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Predicted Alertness",
            "type": "number"
          },
          "raises_situation": {
            "title": "Raises Situation",
            "type": "boolean"
          },
          "risk_level": {
            "enum": [
              "nominal",
              "elevated",
              "critical"
            ],
            "title": "Risk Level",
            "type": "string"
          },
          "scheduled": {
            "format": "date-time",
            "title": "Scheduled",
            "type": "string"
          },
          "situation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Situation Id"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "label",
          "type",
          "criticality",
          "scheduled",
          "assigned_to",
          "assigned_name",
          "predicted_alertness",
          "circadian_low",
          "raises_situation",
          "risk_level"
        ],
        "title": "TimelineTask",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Human Adaptation & Vitality Enhancement Network -- a fatigue-aware safety co-pilot. The deterministic tier owns every safety-critical number; the reasoning tier reads, selects, explains, or refuses; the human owns every decision.",
    "title": "HAVEN",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/audit/{audit_ref}": {
      "get": {
        "operationId": "get_audit_api_audit__audit_ref__get",
        "parameters": [
          {
            "in": "path",
            "name": "audit_ref",
            "required": true,
            "schema": {
              "title": "Audit Ref",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Audit"
      }
    },
    "/api/decisions": {
      "get": {
        "operationId": "list_decisions_api_decisions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DecisionRecord"
                  },
                  "title": "Response List Decisions Api Decisions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Decisions"
      },
      "post": {
        "description": "Stage 7: capture the human decision.\n\nHAVEN never actions anything itself. This endpoint records what the operator\nchose and why, which is also the input to the decision-learning loop.",
        "operationId": "record_decision_api_decisions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Decision"
      }
    },
    "/api/evaluate": {
      "post": {
        "operationId": "post_evaluate_api_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Evaluate"
      }
    },
...
    "/api/scenarios/{scenario_id}/evaluate": {
      "get": {
        "operationId": "evaluate_scenario_api_scenarios__scenario_id__evaluate_get",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Evaluate Scenario"
      }
    }
  }
}